refactor: split synthetic corpus generation stack
Decompose synthetic generation into package selection, runtime emission,
conversation builders, and typed batch/report models while keeping the
SyntheticCorpus facade stable.
