+++ polylogue/schemas/synthetic/__init__.py
@@ -7,14 +7,21 @@ JSON schemas.  All public symbols from the original monolithic
 work unchanged.
 
 Submodules:
-    - ``core`` — SyntheticCorpus class and generation logic
+    - ``core`` — SyntheticCorpus facade and batch/report entrypoints
+    - ``selection`` — package-aware provider/element selection
+    - ``builders`` — conversation builders and wire-format shaping
+    - ``runtime`` — recursive schema-driven value emission
     - ``semantic_values`` — Semantic-role-driven value generation
     - ``relations`` — Relational constraint satisfaction
     - ``wire_formats`` — WireFormat, TreeConfig, PROVIDER_WIRE_FORMATS
     - ``showcase`` — ConversationTheme, _SHOWCASE_THEMES
 """
 
-from polylogue.schemas.synthetic.core import SyntheticCorpus
+from polylogue.schemas.synthetic.core import (
+    SyntheticCorpus,
+    SyntheticGenerationBatch,
+    SyntheticGenerationReport,
+)
 from polylogue.schemas.synthetic.relations import (
     ForeignKeyGraph,
     MutualExclusionGroup,
@@ -23,11 +30,11 @@ from polylogue.schemas.synthetic.relations import (
     TimeDeltaConstraint,
 )
 from polylogue.schemas.synthetic.semantic_values import (
-    SemanticValueGenerator,
     _ROLE_TEXTS,
+    SemanticValueGenerator,
     _text_for_role,
 )
-from polylogue.schemas.synthetic.showcase import ConversationTheme, _SHOWCASE_THEMES
+from polylogue.schemas.synthetic.showcase import _SHOWCASE_THEMES, ConversationTheme
 from polylogue.schemas.synthetic.wire_formats import (
     PROVIDER_WIRE_FORMATS,
     TreeConfig,
@@ -37,6 +44,8 @@ from polylogue.schemas.synthetic.wire_formats import (
 __all__ = [
     # Core
     "SyntheticCorpus",
+    "SyntheticGenerationBatch",
+    "SyntheticGenerationReport",
     # Wire formats
     "PROVIDER_WIRE_FORMATS",
     "TreeConfig",
